refactor: Move moduleTypes to utils/moduleRegistry
- Move from types/ to utils/ directory
- Rename to moduleRegistry.ts for better clarity
- Update all import paths
- Keep same functional API
