+++ packages/frontend/editor-ui/src/utils/moduleRegistry.ts
